Add trick for bin/ and lib/ dirs to show in repo
For reference: http://stackoverflow.com/questions/115983/how-do-i-add-an-empty-directory-to-a-git-repository

The bin/ and lib/ directories are needed by for building the project, but do not show up in the repo since there are no files in these directories that are being tracked. This trick allows these two folders to show up in the repo (and thus appear when the project is cloned).