Change how we read/write input particle states
-- I want to change how Particles are read in by the program. Previously we copied all the initial particle states into the output folder in the Initialisation of Data, and then closed the input file. Then the simulation read each input state from the output file. 

Not only is this a bit inefficient (requiring extra reads from disk), but this causes a problem with holding the initial random number seed in the particle state. Currently we are storing the random number seed in the particle, so that we can restore the state of the random number generator when a particle is reloaded and run again for debugging purposes. However this doesn't work if the initial particle state is saved to disk before it is propagated (as the random number seed can only be saved at the start of propagation). 

To fix this I want to keep the input File/tree/branch in memory and read each particle from the input at the start of propagation, set its initial state, and then write it to the output file  before propagating.